# Fix performance regression in V3D