--- a/Framework/Kernel/inc/MantidKernel/V3D.h
+++ b/Framework/Kernel/inc/MantidKernel/V3D.h
@@ -10,8 +10,11 @@
 #include "MantidKernel/DllConfig.h"
 #include "MantidKernel/Exception.h"
 #include "MantidKernel/Tolerance.h"
+#include <array>
+#include <cassert>
 #include <cmath>
-#include <iosfwd>
+#include <numeric>
+#include <sstream>
 #include <vector>
 
 namespace NeXus {
@@ -30,20 +33,16 @@ Class for 3D vectors.
 */
 class MANTID_KERNEL_DLL V3D final {
 public:
-  constexpr V3D() noexcept : x(0.), y(0.), z(0.) {}
+  constexpr V3D() noexcept : m_pt({{0., 0., 0.}}) {}
   constexpr V3D(double xx, double yy, double zz) noexcept
-      : x(xx), y(yy), z(zz) {}
+      : m_pt({xx, yy, zz}) {}
 
   /// Convenience method for sorting list of V3D objects based on magnitude
-  static bool CompareMagnitude(const Kernel::V3D &v1, const Kernel::V3D &v2);
+  static bool compareMagnitude(const Kernel::V3D &v1, const Kernel::V3D &v2);
 
   // explicit conversion into vector
   operator std::vector<double>() const {
-    std::vector<double> tmp(3);
-    tmp[0] = x;
-    tmp[1] = y;
-    tmp[2] = z;
-    return tmp;
+    return std::vector<double>(m_pt.cbegin(), m_pt.cend());
   }
 
   /**
@@ -52,7 +51,7 @@ public:
      @return *this+v;
   */
   constexpr V3D operator+(const V3D &v) const noexcept {
-    return V3D(x + v.x, y + v.y, z + v.z);
+    return V3D(m_pt[0] + v.m_pt[0], m_pt[1] + v.m_pt[1], m_pt[2] + v.m_pt[2]);
   }
 
   /**
@@ -61,7 +60,7 @@ public:
     @return *this-v;
   */
   constexpr V3D operator-(const V3D &v) const noexcept {
-    return V3D(x - v.x, y - v.y, z - v.z);
+    return V3D(m_pt[0] - v.m_pt[0], m_pt[1] - v.m_pt[1], m_pt[2] - v.m_pt[2]);
   }
 
   /**
@@ -70,7 +69,7 @@ public:
     @return *this * v;
   */
   constexpr V3D operator*(const V3D &v) const noexcept {
-    return V3D(x * v.x, y * v.y, z * v.z);
+    return V3D(m_pt[0] * v.m_pt[0], m_pt[1] * v.m_pt[1], m_pt[2] * v.m_pt[2]);
   }
 
   /**
@@ -79,7 +78,7 @@ public:
     @return *this * v;
   */
   constexpr V3D operator/(const V3D &v) const noexcept {
-    return V3D(x / v.x, y / v.y, z / v.z);
+    return V3D(m_pt[0] / v.m_pt[0], m_pt[1] / v.m_pt[1], m_pt[2] / v.m_pt[2]);
   }
 
   /**
@@ -88,9 +87,9 @@ public:
     @return *this+=v;
   */
   V3D &operator+=(const V3D &v) noexcept {
-    x += v.x;
-    y += v.y;
-    z += v.z;
+    for (size_t i = 0; i < m_pt.size(); ++i) {
+      m_pt[i] += v.m_pt[i];
+    }
     return *this;
   }
 
@@ -100,9 +99,9 @@ public:
     @return *this-v;
   */
   V3D &operator-=(const V3D &v) noexcept {
-    x -= v.x;
-    y -= v.y;
-    z -= v.z;
+    for (size_t i = 0; i < m_pt.size(); ++i) {
+      m_pt[i] -= v.m_pt[i];
+    }
     return *this;
   }
 
@@ -112,9 +111,9 @@ public:
     @return *this*=v;
   */
   V3D &operator*=(const V3D &v) noexcept {
-    x *= v.x;
-    y *= v.y;
-    z *= v.z;
+    for (size_t i = 0; i < m_pt.size(); ++i) {
+      m_pt[i] *= v.m_pt[i];
+    }
     return *this;
   }
 
@@ -124,9 +123,9 @@ public:
     @return *this*=v;
   */
   V3D &operator/=(const V3D &v) noexcept {
-    x /= v.x;
-    y /= v.y;
-    z /= v.z;
+    for (size_t i = 0; i < m_pt.size(); ++i) {
+      m_pt[i] /= v.m_pt[i];
+    }
     return *this;
   }
 
@@ -136,7 +135,7 @@ public:
     @return this * D
    */
   constexpr V3D operator*(const double D) const noexcept {
-    return V3D(x * D, y * D, z * D);
+    return V3D(m_pt[0] * D, m_pt[1] * D, m_pt[2] * D);
   }
 
   /**
@@ -145,7 +144,7 @@ public:
     @return this / D
   */
   constexpr V3D operator/(const double D) const noexcept {
-    return V3D(x / D, y / D, z / D);
+    return V3D(m_pt[0] / D, m_pt[1] / D, m_pt[2] / D);
   }
 
   /**
@@ -154,9 +153,9 @@ public:
     @return this *= D
   */
   V3D &operator*=(const double D) noexcept {
-    x *= D;
-    y *= D;
-    z *= D;
+    for (size_t i = 0; i < m_pt.size(); ++i) {
+      m_pt[i] *= D;
+    }
     return *this;
   }
 
@@ -167,10 +166,8 @@ public:
     \todo ADD TOLERANCE
   */
   V3D &operator/=(const double D) noexcept {
-    if (D != 0.0) {
-      x /= D;
-      y /= D;
-      z /= D;
+    for (size_t i = 0; i < m_pt.size(); ++i) {
+      m_pt[i] /= D;
     }
     return *this;
   }
@@ -179,7 +176,9 @@ public:
     Negation
    * @return a vector with same magnitude but in opposite direction
    */
-  constexpr V3D operator-() const noexcept { return V3D(-x, -y, -z); }
+  constexpr V3D operator-() const noexcept {
+    return V3D(-m_pt[0], -m_pt[1], -m_pt[2]);
+  }
 
   /**
     Equals operator with tolerance factor
@@ -188,8 +187,9 @@ public:
   */
   bool operator==(const V3D &v) const {
     using namespace std;
-    return !(fabs(x - v.x) > Tolerance || fabs(y - v.y) > Tolerance ||
-             fabs(z - v.z) > Tolerance);
+    return !(fabs(m_pt[0] - v.m_pt[0]) > Tolerance ||
+             fabs(m_pt[1] - v.m_pt[1]) > Tolerance ||
+             fabs(m_pt[2] - v.m_pt[2]) > Tolerance);
   }
 
   /** Not equals operator with tolerance factor.
@@ -203,11 +203,11 @@ public:
     @return true if V is greater
    */
   constexpr bool operator<(const V3D &V) const noexcept {
-    if (x != V.x)
-      return x < V.x;
-    if (y != V.y)
-      return y < V.y;
-    return z < V.z;
+    if (m_pt[0] != V.m_pt[0])
+      return m_pt[0] < V.m_pt[0];
+    if (m_pt[1] != V.m_pt[1])
+      return m_pt[1] < V.m_pt[1];
+    return m_pt[2] < V.m_pt[2];
   }
 
   /// Comparison operator greater than.
@@ -222,56 +222,45 @@ public:
     @param zz :: The Z coordinate
   */
   void operator()(const double xx, const double yy, const double zz) noexcept {
-    x = xx;
-    y = yy;
-    z = zz;
+    m_pt = {{xx, yy, zz}};
   }
 
   // Access
   // Setting x, y and z values
-  void spherical(const double &R, const double &theta, const double &phi);
-  void spherical_rad(const double &R, const double &polar,
-                     const double &azimuth);
-  void azimuth_polar_SNS(const double &R, const double &azimuth,
-                         const double &polar);
+  void spherical(const double R, const double theta, const double phi);
+  void spherical_rad(const double R, const double polar, const double azimuth);
+  void azimuth_polar_SNS(const double R, const double azimuth,
+                         const double polar);
   /**
     Set is x position
     @param xx :: The X coordinate
   */
-  void setX(const double xx) noexcept { x = xx; }
+  void setX(const double xx) noexcept { m_pt[0] = xx; }
 
   /**
     Set is y position
     @param yy :: The Y coordinate
   */
-  void setY(const double yy) noexcept { y = yy; }
+  void setY(const double yy) noexcept { m_pt[1] = yy; }
 
   /**
     Set is z position
     @param zz :: The Z coordinate
   */
-  void setZ(const double zz) noexcept { z = zz; }
+  void setZ(const double zz) noexcept { m_pt[2] = zz; }
 
-  const double &X() const noexcept { return x; } ///< Get x
-  const double &Y() const noexcept { return y; } ///< Get y
-  const double &Z() const noexcept { return z; } ///< Get z
+  constexpr double X() const noexcept { return m_pt[0]; } ///< Get x
+  constexpr double Y() const noexcept { return m_pt[1]; } ///< Get y
+  constexpr double Z() const noexcept { return m_pt[2]; } ///< Get z
 
   /**
     Returns the axis value based in the index provided
-    @param Index :: 0=x, 1=y, 2=z
+    @param index :: 0=x, 1=y, 2=z
     @return a double value of the requested axis
   */
-  const double &operator[](const size_t Index) const {
-    switch (Index) {
-    case 0:
-      return x;
-    case 1:
-      return y;
-    case 2:
-      return z;
-    default:
-      throw Kernel::Exception::IndexError(Index, 2, "operator[] range error");
-    }
+  constexpr double operator[](const size_t index) const {
+    assert(index < m_pt.size());
+    return m_pt[index];
   }
 
   /**
@@ -279,17 +268,9 @@ public:
     @param Index :: 0=x, 1=y, 2=z
     @return a double value of the requested axis
   */
-  double &operator[](const size_t Index) {
-    switch (Index) {
-    case 0:
-      return x;
-    case 1:
-      return y;
-    case 2:
-      return z;
-    default:
-      throw Kernel::Exception::IndexError(Index, 2, "operator[] range error");
-    }
+  double &operator[](const size_t index) {
+    assert(index < m_pt.size());
+    return m_pt[index];
   }
 
   void getSpherical(double &R, double &theta, double &phi) const;
@@ -297,15 +278,13 @@ public:
   void rotate(const Matrix<double> &);
 
   void round();
-
   /// Make a normalized vector (return norm value)
-  double normalize(); // Vec3D::makeUnit
+  double normalize();
   double norm() const;
-  /**
-    Vector length without the sqrt
-    @return vec.length()
-  */
-  constexpr double norm2() const noexcept { return x * x + y * y + z * z; }
+  /// Vector length squared
+  constexpr double norm2() const noexcept {
+    return m_pt[0] * m_pt[0] + m_pt[1] * m_pt[1] + m_pt[2] * m_pt[2];
+  }
   /// transform vector into form, used to describe directions in
   /// crystallogaphical coodinate system
   double toMillerIndexes(double eps = 1.e-3);
@@ -315,14 +294,20 @@ public:
     @return The cross product of the two vectors (this * v)
   */
   constexpr double scalar_prod(const V3D &v) const noexcept {
-    return x * v.x + y * v.y + z * v.z;
+    return m_pt[0] * v.m_pt[0] + m_pt[1] * v.m_pt[1] + m_pt[2] * v.m_pt[2];
   }
   /// Cross product (this * argument)
   constexpr V3D cross_prod(const V3D &v) const noexcept {
-    return V3D(y * v.z - z * v.y, z * v.x - x * v.z, x * v.y - y * v.x);
+    return V3D(m_pt[1] * v.m_pt[2] - m_pt[2] * v.m_pt[1],
+               m_pt[2] * v.m_pt[0] - m_pt[0] * v.m_pt[2],
+               m_pt[0] * v.m_pt[1] - m_pt[1] * v.m_pt[0]);
   }
-  /// Distance (R) between two points defined as vectors
-  double distance(const V3D &) const;
+  /**
+    Calculates the distance between two vectors
+    @param v :: The second vector to include in the calculation
+    @return The distance between the two vectors
+  */
+  double distance(const V3D &v) const { return (*this - v).norm(); }
   /// Zenith (theta) angle between this and another vector
   double zenith(const V3D &) const;
   /// Angle between this and another vector
@@ -341,25 +326,21 @@ public:
   std::string toString() const;
   void fromString(const std::string &str);
 
-  double volume() const {
-    return fabs(x * y * z);
-  } ///< Calculate the volume of a cube X*Y*Z
-
-  int reBase(const V3D &, const V3D &,
-             const V3D &); ///< rebase to new basis vector
-  int masterDir(const double Tol =
-                    1e-3) const; ///< Determine if there is a master direction
-  bool
-  nullVector(const double Tol = 1e-3) const; ///< Determine if the point is null
+  /// Calculate the volume of a cube X*Y*Z
+  double volume() const { return fabs(m_pt[0] * m_pt[1] * m_pt[2]); }
+  /// rebase to new basis vector
+  int reBase(const V3D &, const V3D &, const V3D &);
+  /// Determine if there is a master direction
+  int masterDir(const double Tol = 1e-3) const;
+  /// Determine if the point is null
+  bool nullVector(const double tolerance = 1e-3) const;
   bool coLinear(const V3D &, const V3D &) const;
 
   void saveNexus(::NeXus::File *file, const std::string &name) const;
   void loadNexus(::NeXus::File *file, const std::string &name);
 
 private:
-  double x; ///< X value [unitless]
-  double y; ///< Y value [unitless]
-  double z; ///< Z value [unitless]
+  std::array<double, 3> m_pt;
 };
 
 // Overload operator <<