--- a/Framework/PythonInterface/mantid/kernel/src/Exports/V3D.cpp
+++ b/Framework/PythonInterface/mantid/kernel/src/Exports/V3D.cpp
@@ -101,22 +101,16 @@ void export_V3D() {
       .def(init<double, double, double>(
           "Construct a V3D with X,Y,Z coordinates"))
       .def("X", &V3D::X, arg("self"),
-           return_value_policy<copy_const_reference>(),
            "Returns the X coordinate")
       .def("Y", &V3D::Y, arg("self"),
-           return_value_policy<copy_const_reference>(),
            "Returns the Y coordinate")
       .def("Z", &V3D::Z, arg("self"),
-           return_value_policy<copy_const_reference>(),
            "Returns the Z coordinate")
       .def("getX", &V3D::X, arg("self"),
-           return_value_policy<copy_const_reference>(),
            "Returns the X coordinate") // Traditional name
       .def("getY", &V3D::Y, arg("self"),
-           return_value_policy<copy_const_reference>(),
            "Returns the Y coordinate") // Traditional name
       .def("getZ", &V3D::Z, arg("self"),
-           return_value_policy<copy_const_reference>(),
            "Returns the Z coordinate") // Traditional name
       .def("distance", &V3D::distance, (arg("self"), arg("other")),
            "Returns the distance between this vector and another")