… Framework/Kernel/test/V3DTest.h
… qt/widgets/instrumentview/src/InstrumentActor.cpp
… qt/widgets/instrumentview/src/UnwrappedSurface.cpp