@DanNixon 2019-03-12
Unfortunately this has caused another regression in [KernelTest.V3DTestPerformance.testAddAssignOperatorV3D](http://builds.mantidproject.org/job/master_performancetests2/Master_20branch_20performance_20tests/KernelTest.V3DTestPerformance.testAddAssignOperatorV3D.htm).